APM-1216 split up script to seperate class
